+++ src/Ambev.DeveloperEvaluation.Domain/Common/PaginatedList.cs
@@ -4,10 +4,10 @@ namespace Ambev.DeveloperEvaluation.Domain.Common;
 
 public class PaginatedList<T> : List<T>
 {
-    public int CurrentPage { get; private set; }
-    public int TotalPages { get; private set; }
-    public int PageSize { get; private set; }
-    public int TotalItems { get; private set; }
+    public int CurrentPage { get; }
+    public int TotalPages { get; }
+    public int PageSize { get; }
+    public int TotalItems { get; }
 
     public bool HasPrevious => CurrentPage > 1;
     public bool HasNext => CurrentPage < TotalPages;
